+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2019-02-15
+
+### Added
+- Support for cancelling certificate update. This can occur when a stack update is cancelled, perhaps due to another
+  resource failing to create/update/delete.
+
 ## [1.6.0] - 2019-02-01
 
 ### Added
@@ -51,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First release
 
+[1.7.0]: https://github.com/dflook/cloudformation-dns-certificate/compare/1.6.0...1.7.0
 [1.6.0]: https://github.com/dflook/cloudformation-dns-certificate/compare/1.5.1...1.6.0
 [1.5.1]: https://github.com/dflook/cloudformation-dns-certificate/compare/1.5.0...1.5.1
 [1.5.0]: https://github.com/dflook/cloudformation-dns-certificate/compare/a64051e43ae8696c898b6634fbe663abc4a87785...1.5.0